requirements.txt
flask
mysql-connector-python
python-dotenv
requests
openai
flask-cors
google-generativeai
Werkzeug
cohere
redis
